+++ app/server/lib/petitions.js
@@ -1,0 +1,10 @@
+import { runQuery } from './sql.js';
+
+export async function getPetitions(app) {
+    return runQuery(
+        app.db,
+        `
+        SELECT * FROM petitions
+        ;`
+    );
+}
