+++ app/server/lib/sql.js
@@ -1,14 +1,14 @@
-const fs = require('fs');
-const zipObject = require('lodash/zipObject');
-const initSqlJs = require('sql.js');
+import fs from 'fs';
+import zipObject from 'lodash/zipObject.js';
+import initSqlJs from 'sql.js';
 
-async function getConnection(file) {
+export async function getConnection(file) {
     const fileBuffer = fs.readFileSync(file);
     const SQL = await initSqlJs();
     return new SQL.Database(fileBuffer);
 }
 
-async function runQuery(db, sqlStatement) {
+export async function runQuery(db, sqlStatement) {
     const response = db.exec(sqlStatement);
 
     if (response.length) {
@@ -18,8 +18,3 @@ async function runQuery(db, sqlStatement) {
         return [];
     }
 }
-
-module.exports = {
-    getConnection,
-    runQuery
-};
