--- app/server/routes/health.js
@@ -1,9 +1,0 @@
-const jsonPromiseHandler = require('../lib/jsonPromiseHandler');
-
-async function healthcheck(req) {
-    return 'ALIVE';
-}
-
-module.exports = (app) => {
-    app.get('/health', jsonPromiseHandler(healthcheck));
-};
