--- config.json
@@ -1,3 +1,0 @@
-{
-    "port": 3000
-}
