+++ support/testing/index.js
@@ -1,10 +1,3 @@
-const eventually = require('./eventually');
-const { Page } = require('./page');
-const { seedDbWithPetitionCount, savePetition } = require('./seeding');
-
-module.exports = {
-    Page,
-    eventually,
-    seedDbWithPetitionCount,
-    savePetition
-};
+export * from './eventually';
+export * from './page';
+export * from './seeding';
