--- support/testing/oneliner.js
@@ -1,5 +1,0 @@
-const oneliner = (text) => text.replace(/\s+|\n/g, ' ').trim();
-
-module.exports = {
-    oneliner
-};
